requirements.txt
asgiref==3.3.1
autopep8==1.5.4
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cryptography==3.3.1
defusedxml==0.6.0
Django==3.1.4
django-allauth==0.44.0
django-cors-headers==3.6.0
django-countries==7.0
django-crispy-forms==1.10.0
django-debug-toolbar==3.2
django-environ==0.4.5
django-localflavor==3.0.1
django-multiselectfield==0.1.12
django-rest-passwordreset==1.1.0
djangorestframework==3.12.2
idna==2.10
oauthlib==3.1.0
pep8==1.7.1
Pillow==8.0.1
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
python-decouple==3.3
python-stdnum==1.16
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
stripe==2.55.1
toml==0.10.2
urllib3==1.26.2
whitenoise==5.2.0
